requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
distlib==0.3.8
filelock==3.13.1
Flask==3.0.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
network==0.1
platformdirs==4.2.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
soupsieve==2.5
urllib3==2.0.6
virtualenv==20.25.1
Werkzeug==3.0.3
